feat(auth): Implement face enrollment
This commit introduces the functionality for face enrollment. It adds the necessary dependencies, components, and UI for capturing and enrolling user faces.

- Add `face-api.js`, `@simplewebauthn/browser`, and `react-webcam` dependencies.
- Create `FaceEnrollment` component for face capture.
- Integrate `FaceEnrollment` component into the students and welcome pages.
- Update API base URL to production.

+++ src/app/page.tsx
@@ -9,12 +9,10 @@ const HomePage = () => {
 
   useEffect(() => {
     const teachersEmail = session?.user?.email;
-    localStorage.setItem("email", teachersEmail as string);
-    async function biometricCheck() {
-      const capabilities = await PublicKeyCredential.getClientCapabilities();
-      console.log(capabilities);
+    const emailExists = localStorage.getItem("email");
+    if (!emailExists || emailExists === "undefined") {
+      localStorage.setItem("email", teachersEmail as string);
     }
-    biometricCheck();
   }, [session, status]);
 
   if (session === undefined) return null;
